Move charts into separate directories and files
Just separating the code out a bit to make it a little more accessible.
