+++ packages/transform-dataresource/src/charts.js
@@ -1,863 +1,19 @@
 /* @flow */
-import * as React from "react";
 
-import { nest } from "d3-collection";
-import { scaleLinear, scaleTime, scaleThreshold } from "d3-scale";
 import {
   ResponsiveOrdinalFrame,
   ResponsiveXYFrame,
   ResponsiveNetworkFrame
 } from "semiotic";
-import HTMLLegend from "./HTMLLegend";
+
 import ParallelCoordinatesController from "./ParallelCoordinatesController";
-import { numeralFormatting } from "./utilities";
-import { curveMonotone } from "d3-shape";
 
-function combineTopAnnotations(
-  topQ: Array<Object>,
-  topSecondQ: Array<Object>,
-  dim2: string
-): any[] {
-  const combinedAnnotations = [];
-  const combinedHash = {};
-  [...topQ, ...topSecondQ].forEach(d => {
-    const hashD = combinedHash[d[dim2]];
-
-    if (hashD) {
-      const newCoordinates = (hashD.coordinates && [
-        ...hashD.coordinates,
-        d
-      ]) || [d, hashD];
-      Object.keys(combinedHash[d[dim2]]).forEach(k => {
-        delete combinedHash[d[dim2]][k];
-      });
-      combinedHash[d[dim2]].id = d[dim2];
-      combinedHash[d[dim2]].label = d[dim2];
-      combinedHash[d[dim2]].type = "react-annotation";
-      combinedHash[d[dim2]].coordinates = newCoordinates;
-    } else {
-      combinedHash[d[dim2]] = {
-        type: "react-annotation",
-        label: d[dim2],
-        id: d[dim2],
-        ...d
-      };
-      combinedAnnotations.push(combinedHash[d[dim2]]);
-    }
-  });
-  return combinedAnnotations;
-}
-
-function stringOrFnAccessor(d: Object, accessor: string | Function) {
-  return typeof accessor === "function" ? accessor(d) : d[accessor];
-}
-
-function sortByOrdinalRange(
-  oAccessor: Function | string,
-  rAccessor: Function | string,
-  secondarySort: string,
-  data: Array<Object>
-): any[] {
-  const subsortData = {};
-  let subsortArrays = [];
-  data.forEach(d => {
-    const oD = stringOrFnAccessor(d, oAccessor);
-    if (!subsortData[oD]) {
-      subsortData[oD] = { array: [], value: 0, label: oD };
-      subsortArrays.push(subsortData[oD]);
-    }
-    subsortData[oD].array.push(d);
-    subsortData[oD].value += stringOrFnAccessor(d, rAccessor);
-  });
-
-  subsortArrays = subsortArrays.sort((a, b) => {
-    if (b.value === a.value) {
-      if (a.label < b.label) return -1;
-      if (a.label > b.label) return 1;
-      return 1;
-    }
-
-    return b.value - a.value;
-  });
-
-  if (secondarySort !== "none") {
-    subsortArrays.forEach(a => {
-      a.array = a.array.sort(
-        (a, b) =>
-          stringOrFnAccessor(b, secondarySort) -
-          stringOrFnAccessor(a, secondarySort)
-      );
-    });
-  }
-
-  return subsortArrays.reduce((p, c) => [...p, ...c.array], []);
-}
-
-const steps = ["none", "#FBEEEC", "#f3c8c2", "#e39787", "#ce6751", "#b3331d"];
-const thresholds = scaleThreshold()
-  .domain([0.01, 0.2, 0.4, 0.6, 0.8])
-  .range(steps);
-
-const parentPath = (d, pathArray) => {
-  if (d.parent) {
-    pathArray = parentPath(d.parent, [d.key, ...pathArray]);
-  } else {
-    pathArray = ["root", ...pathArray];
-  }
-  return pathArray;
-};
-
-const fontScale = scaleLinear()
-  .domain([8, 25])
-  .range([14, 8])
-  .clamp(true);
-
-const hierarchicalTooltip = (d, primaryKey, metric) => {
-  const pathString = d.parent
-    ? parentPath(d.parent, (d.key && [d.key]) || []).join("->")
-    : "";
-  const content = [];
-  if (!d.parent) {
-    content.push(<h2 key="hierarchy-title">Root</h2>);
-  } else if (d.key) {
-    content.push(<h2 key="hierarchy-title">{d.key}</h2>);
-    content.push(<p key="path-string">{pathString}</p>);
-    content.push(<p key="hierarchy-value">Total Value: {d.value}</p>);
-    content.push(<p key="hierarchy-children">Children: {d.children.length}</p>);
-  } else {
-    content.push(
-      <p key="leaf-label">
-        {pathString}->{primaryKey.map(p => d[p]).join(", ")}
-      </p>
-    );
-    content.push(
-      <p key="hierarchy-value">
-        {metric}: {d[metric]}
-      </p>
-    );
-  }
-
-  return content;
-};
-
-const hierarchicalColor = (colorHash: Object, d: Object) => {
-  if (d.depth === 0) return "white";
-  if (d.depth === 1) return colorHash[d.key];
-  let colorNode = d;
-  for (let x = d.depth; x > 1; x--) {
-    colorNode = colorNode.parent;
-  }
-  const lightenScale = scaleLinear()
-    .domain([6, 1])
-    .clamp(true)
-    .range(["white", colorHash[colorNode.key]]);
-
-  return lightenScale(d.depth);
-};
-
-const semioticLineChart = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  let lineData;
-
-  const {
-    chart,
-    selectedMetrics,
-    lineType,
-    metrics,
-    dimensions,
-    primaryKey,
-    colors
-  } = options;
-
-  const { timeseriesSort } = chart;
-
-  const sortType =
-    timeseriesSort === "array-order"
-      ? "integer"
-      : schema.fields.find(p => p.name === timeseriesSort).type;
-
-  const formatting =
-    sortType === "datetime"
-      ? d => d.toLocaleString().split(",")[0]
-      : numeralFormatting;
-
-  const xScale = sortType === "datetime" ? scaleTime() : scaleLinear();
-
-  lineData = metrics
-    .map((d, i) => {
-      const metricData =
-        timeseriesSort === "array-order"
-          ? data
-          : data.sort((a, b) => a[timeseriesSort] - b[timeseriesSort]);
-      return {
-        color: colors[i % colors.length],
-        label: d.name,
-        type: d.type,
-        coordinates: metricData.map((p, q) => ({
-          value: p[d.name],
-          x: timeseriesSort === "array-order" ? q : p[timeseriesSort],
-          label: d.name,
-          color: colors[i % colors.length],
-          originalData: p
-        }))
-      };
-    })
-    .filter(
-      d =>
-        selectedMetrics.length === 0 || selectedMetrics.find(p => p === d.label)
-    );
-
-  return {
-    lineType: { type: lineType, interpolator: curveMonotone },
-    lines: lineData,
-    xScaleType: xScale,
-    renderKey: (d: Object, i: number) => {
-      return d.coordinates ? `line-${d.label}` : `linepoint=${d.label}-${i}`;
-    },
-    lineStyle: (d: Object) => ({
-      fill: lineType === "line" ? "none" : d.color,
-      stroke: d.color,
-      fillOpacity: 0.75
-    }),
-    pointStyle: (d: Object) => {
-      return {
-        fill: d.color,
-        fillOpacity: 0.75
-      };
-    },
-    axes: [
-      { orient: "left", tickFormat: numeralFormatting },
-      {
-        orient: "bottom",
-        ticks: 5,
-        tickFormat: (d: any) => {
-          const label = formatting(d);
-          const rotation = label.length > 4 ? "45" : "0";
-          const textAnchor = label.length > 4 ? "start" : "middle";
-          return (
-            <text transform={`rotate(${rotation})`} textAnchor={textAnchor}>
-              {label}
-            </text>
-          );
-        }
-      }
-    ],
-    hoverAnnotation: true,
-    xAccessor: "x",
-    yAccessor: "value",
-    showLinePoints: lineType === "line",
-    margin: {
-      top: 20,
-      right: 200,
-      bottom: sortType === "datetime" ? 80 : 40,
-      left: 50
-    },
-    legend: {
-      title: "Legend",
-      position: "right",
-      width: 200,
-      legendGroups: [
-        {
-          label: "",
-          styleFn: (d: Object) => ({ fill: d.color }),
-          items: lineData
-        }
-      ]
-    },
-    tooltipContent: (d: Object) => {
-      return (
-        <div className="tooltip-content">
-          <p>{d.parentLine && d.parentLine.label}</p>
-          <p>
-            {d.label}:{" "}
-            {(d.value.toLocaleString && d.value.toLocaleString()) || d.value}
-          </p>
-          <p>{formatting(d.x)}</p>
-          {primaryKey.map((k, ki) => (
-            <p key={`key-${ki}`}>{d.originalData[k]}</p>
-          ))}
-        </div>
-      );
-    }
-  };
-};
-
-const semioticNetwork = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  const { networkType = "force", chart, colors } = options;
-  const { dim1: sourceDimension, dim2: targetDimension, metric1 } = chart;
-
-  if (
-    !sourceDimension ||
-    sourceDimension === "none" ||
-    !targetDimension ||
-    targetDimension === "none"
-  ) {
-    return {};
-  }
-  const edgeHash = {};
-  const networkData = [];
-
-  data.forEach(d => {
-    if (!edgeHash[`${d[sourceDimension]}-${d[targetDimension]}`]) {
-      edgeHash[`${d[sourceDimension]}-${d[targetDimension]}`] = {
-        source: d[sourceDimension],
-        target: d[targetDimension],
-        value: 0,
-        weight: 0
-      };
-      networkData.push(edgeHash[`${d[sourceDimension]}-${d[targetDimension]}`]);
-    }
-    edgeHash[`${d[sourceDimension]}-${d[targetDimension]}`].value +=
-      d[metric1] || 1;
-    edgeHash[`${d[sourceDimension]}-${d[targetDimension]}`].weight += 1;
-  });
-
-  const colorHash = {};
-  data.forEach(d => {
-    if (!colorHash[d[sourceDimension]])
-      colorHash[d[sourceDimension]] =
-        colors[Object.keys(colorHash).length % colors.length];
-    if (!colorHash[d[targetDimension]])
-      colorHash[d[targetDimension]] =
-        colors[Object.keys(colorHash).length % colors.length];
-  });
-
-  networkData.forEach(d => {
-    d.weight = Math.min(10, d.weight);
-  });
-
-  return {
-    edges: networkData,
-    edgeType: "halfarrow",
-    edgeStyle: (d: Object) => ({
-      fill: colorHash[d.source.id],
-      stroke: colorHash[d.source.id],
-      strokeOpacity: 0.5
-    }),
-    nodeStyle: (d: Object) => ({
-      fill: colorHash[d.id],
-      stroke: colorHash[d.id],
-      strokeOpacity: 0.5
-    }),
-    nodeSizeAccessor: (d: Object) => d.degree,
-    networkType: {
-      type: networkType,
-      iterations: 1000
-    },
-    hoverAnnotation: true,
-    tooltipContent: (d: Object) => {
-      return (
-        <div className="tooltip-content">
-          <h2>{d.id}</h2>
-          <p>Links: {d.degree}</p>
-          {d.value && <p>Value: {d.value}</p>}
-        </div>
-      );
-    },
-    margin: { left: 100, right: 100, top: 10, bottom: 10 }
-  };
-};
-
-const semioticHierarchicalChart = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  const {
-    hierarchyType = "dendrogram",
-    chart,
-    selectedDimensions,
-    primaryKey,
-    colors
-  } = options;
-  const { metric1 } = chart;
-
-  if (selectedDimensions.length === 0) {
-    return {};
-  }
-
-  const nestingParams = nest();
-
-  selectedDimensions.forEach(d => {
-    nestingParams.key(p => p[d]);
-  });
-  const colorHash = {};
-  const sanitizedData = [];
-  data.forEach(d => {
-    if (!colorHash[d[selectedDimensions[0]]])
-      colorHash[d[selectedDimensions[0]]] =
-        colors[Object.keys(colorHash).length];
-
-    sanitizedData.push({
-      ...d,
-      sanitizedR: d.r,
-      r: undefined
-    });
-  });
-
-  const entries = nestingParams.entries(sanitizedData);
-  const rootNode = { values: entries };
-
-  return {
-    edges: rootNode,
-    edgeStyle: (d: Object) => ({ fill: "lightgray", stroke: "gray" }),
-    nodeStyle: (d: Object) => {
-      return {
-        fill: hierarchicalColor(colorHash, d),
-        stroke: d.depth === 1 ? "white" : "black",
-        strokeOpacity: d.depth * 0.1 + 0.2
-      };
-    },
-    networkType: {
-      type: hierarchyType,
-      hierarchySum: (d: Object) => d[metric1],
-      hierarchyChildren: (d: Object) => d.values,
-      padding:
-        hierarchyType === "treemap" ? 3 : hierarchyType === "circlepack" ? 2 : 0
-    },
-    edgeRenderKey: (d: Object, i: number) => {
-      return i;
-    },
-    baseMarkProps: { forceUpdate: true },
-    margin: { left: 100, right: 100, top: 10, bottom: 10 },
-    hoverAnnotation: true,
-    tooltipContent: (d: Object) => {
-      return (
-        <div className="tooltip-content">
-          {hierarchicalTooltip(d, primaryKey, metric1)}
-        </div>
-      );
-    }
-  };
-};
-
-const semioticBarChart = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  const {
-    selectedMetrics,
-    selectedDimensions,
-    chart,
-    colors,
-    setColor
-  } = options;
-  const { dim1, dim2, metric1, metric3 } = chart;
-
-  const oAccessor =
-    selectedDimensions.length === 0
-      ? dim1
-      : (d: Object) => selectedDimensions.map(p => d[p]).join(",");
-
-  const rAccessor = metric1;
-
-  const additionalSettings = {};
-  const colorHash = { Other: "grey" };
-
-  const sortedData = sortByOrdinalRange(
-    oAccessor,
-    (metric3 !== "none" && metric3) || rAccessor,
-    dim1,
-    data
-  );
-
-  if (metric3 && metric3 !== "none") {
-    additionalSettings.dynamicColumnWidth = metric3;
-  }
-
-  if (dim1 && dim1 !== "none") {
-    const uniqueValues = sortedData.reduce(
-      (p, c) =>
-        (!p.find(d => d === c[dim1].toString()) && [
-          ...p,
-          c[dim1].toString()
-        ]) ||
-        p,
-      []
-    );
-
-    uniqueValues.forEach((d: string, i: number) => {
-      colorHash[d] = i > 18 ? "grey" : colors[i % colors.length];
-    });
-
-    additionalSettings.afterElements = (
-      <HTMLLegend
-        valueHash={{}}
-        values={uniqueValues}
-        colorHash={colorHash}
-        setColor={setColor}
-        colors={colors}
-      />
-    );
-
-    if (
-      selectedDimensions.length > 0 &&
-      selectedDimensions.join(",") !== dim1
-    ) {
-      additionalSettings.pieceHoverAnnotation = true;
-      additionalSettings.tooltipContent = d => {
-        return (
-          <div className="tooltip-content">
-            {dim1 && dim1 !== "none" && <p>{d[dim1]}</p>}
-            <p>
-              {typeof oAccessor === "function" ? oAccessor(d) : d[oAccessor]}
-            </p>
-            <p>
-              {rAccessor}: {d[rAccessor]}
-            </p>
-            {metric3 &&
-              metric3 !== "none" && (
-                <p>
-                  {metric3}: {d[metric3]}
-                </p>
-              )}
-          </div>
-        );
-      };
-    }
-  }
-
-  const barSettings = {
-    type: "bar",
-    data: sortedData,
-    oAccessor,
-    rAccessor,
-    style: (d: Object) => ({
-      fill: colorHash[d[dim1]] || colors[0],
-      stroke: colorHash[d[dim1]] || colors[0]
-    }),
-    oPadding: 5,
-    oLabel: (d: Object) => {
-      return <text transform="rotate(90)">{d}</text>;
-    },
-    hoverAnnotation: true,
-    margin: { top: 10, right: 10, bottom: 100, left: 70 },
-    axis: {
-      orient: "left",
-      label: rAccessor,
-      tickFormat: numeralFormatting
-    },
-    tooltipContent: (d: Object) => {
-      return (
-        <div className="tooltip-content">
-          <p>
-            {typeof oAccessor === "function"
-              ? oAccessor(d.pieces[0])
-              : d.pieces[0][oAccessor]}
-          </p>
-          <p>
-            {rAccessor}:{" "}
-            {d.pieces.map(p => p[rAccessor]).reduce((p, c) => p + c, 0)}
-          </p>
-          {metric3 &&
-            metric3 !== "none" && (
-              <p>
-                {metric3}:{" "}
-                {d.pieces.map(p => p[metric3]).reduce((p, c) => p + c, 0)}
-              </p>
-            )}
-        </div>
-      );
-    },
-    baseMarkProps: { forceUpdate: true },
-    ...additionalSettings
-  };
-
-  return barSettings;
-};
-
-const semioticSummaryChart = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  const additionalSettings = {};
-  const colorHash = {};
-
-  const { chart, summaryType, primaryKey, colors, setColor } = options;
-
-  const { dim1, metric1 } = chart;
-
-  const oAccessor = dim1;
-
-  const rAccessor = metric1;
-
-  if (dim1 && dim1 !== "none") {
-    const uniqueValues = data.reduce(
-      (p, c) =>
-        (!p.find(d => d === c[dim1].toString()) && [
-          ...p,
-          c[dim1].toString()
-        ]) ||
-        p,
-      []
-    );
-
-    uniqueValues.forEach((d, i) => {
-      colorHash[d] = colors[i % colors.length];
-    });
-
-    additionalSettings.afterElements = (
-      <HTMLLegend
-        valueHash={{}}
-        values={uniqueValues}
-        colorHash={colorHash}
-        setColor={setColor}
-        colors={colors}
-      />
-    );
-  }
-
-  const summarySettings = {
-    summaryType: { type: summaryType, bins: 16, amplitude: 20 },
-    type: summaryType === "violin" && "swarm",
-    projection: "horizontal",
-    data: data,
-    oAccessor,
-    rAccessor,
-    summaryStyle: (d: Object) => ({
-      fill: colorHash[d[dim1]] || colors[0],
-      fillOpacity: 0.8,
-      stroke: colorHash[d[dim1]] || colors[0]
-    }),
-    style: (d: Object) => ({
-      fill: colorHash[d[dim1]] || colors[0],
-      stroke: "white"
-    }),
-    oPadding: 5,
-    oLabel: (d: Object) => (
-      <text textAnchor="end" fontSize={`${(d && fontScale(d.length)) || 12}px`}>
-        {d}
-      </text>
-    ),
-    margin: { top: 25, right: 10, bottom: 50, left: 100 },
-    axis: {
-      orient: "left",
-      label: rAccessor,
-      tickFormat: numeralFormatting
-    },
-    baseMarkProps: { forceUpdate: true },
-    pieceHoverAnnotation: summaryType === "violin",
-    tooltipContent: (d: Object) => (
-      <div className="tooltip-content">
-        <h2>{primaryKey.map(p => d[p]).join(", ")}</h2>
-        <p>
-          {dim1}: {d[dim1]}
-        </p>
-        <p>
-          {rAccessor}: {d[rAccessor]}
-        </p>
-      </div>
-    ),
-    ...additionalSettings
-  };
-
-  return summarySettings;
-};
-
-const semioticHexbin = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object
-) => {
-  return semioticScatterplot(data, schema, options, true);
-};
-
-const semioticScatterplot = (
-  data: Array<Object>,
-  schema: Object,
-  options: Object,
-  hexbin: boolean = false
-) => {
-  const height = options.height - 150 || 500;
-
-  const { chart, primaryKey, colors, setColor } = options;
-
-  const { dim1, dim2, metric1, metric2, metric3 } = chart;
-
-  const pointTooltip = (d: Object) => (
-    <div className="tooltip-content">
-      <h2>{primaryKey.map(p => d[p]).join(", ")}</h2>
-      {dim1 &&
-        dim1 !== "none" && (
-          <p>
-            {dim1}: {d[dim1]}
-          </p>
-        )}
-      <p>
-        {metric1}: {d[metric1]}
-      </p>
-      <p>
-        {metric2}: {d[metric2]}
-      </p>
-      {metric3 &&
-        metric3 !== "none" && (
-          <p>
-            {metric3}: {d[metric3]}
-          </p>
-        )}
-    </div>
-  );
-
-  const areaTooltip = (d: Object) => (
-    <div className="tooltip-content">
-      <h2
-        style={{
-          fontSize: "14px",
-          textTransform: "uppercase",
-          margin: "5px",
-          fontWeight: 900
-        }}
-      >
-        ID, {metric1}, {metric2}
-      </h2>
-      {d.binItems.map(d => (
-        <p
-          style={{
-            fontSize: "12px",
-            textTransform: "uppercase",
-            margin: "5px"
-          }}
-        >
-          {primaryKey.map(p => d[p]).join(", ")}, {d[metric1]}, {d[metric2]}
-        </p>
-      ))}
-    </div>
-  );
-
-  let sizeScale = e => 5;
-  const colorHash = { Other: "grey" };
-  const additionalSettings = {};
-
-  let annotations;
-
-  if (dim2 && dim2 !== "none") {
-    const topQ = [...data]
-      .sort((a, b) => b[metric1] - a[metric1])
-      .filter((d, i) => i < 3);
-    const topSecondQ = [...data]
-      .sort((a, b) => b[metric2] - a[metric2])
-      .filter(d => topQ.indexOf(d) === -1)
-      .filter((d, i) => i < 3);
-
-    annotations = combineTopAnnotations(topQ, topSecondQ, dim2);
-  }
-
-  if (metric3 && metric3 !== "none") {
-    const dataMin = Math.min(...data.map(d => d[metric3]));
-    const dataMax = Math.max(...data.map(d => d[metric3]));
-    sizeScale = scaleLinear()
-      .domain([dataMin, dataMax])
-      .range([2, 20]);
-  }
-  const sortedData = sortByOrdinalRange(
-    metric1,
-    (metric3 !== "none" && metric3) || metric2,
-    "none",
-    data
-  );
-
-  if (!hexbin && dim1 && dim1 !== "none") {
-    const uniqueValues = sortedData.reduce(
-      (p, c) =>
-        (!p.find(d => d === c[dim1].toString()) && [
-          ...p,
-          c[dim1].toString()
-        ]) ||
-        p,
-      []
-    );
-
-    uniqueValues.forEach((d, i) => {
-      colorHash[d] = i > 18 ? "grey" : colors[i % colors.length];
-    });
-
-    additionalSettings.afterElements = (
-      <HTMLLegend
-        valueHash={{}}
-        values={uniqueValues}
-        colorHash={colorHash}
-        setColor={setColor}
-        colors={colors}
-      />
-    );
-  } else if (hexbin) {
-    const hexValues = [
-      "0% - 20%",
-      "20% - 40%",
-      "40% - 60%",
-      "60% - 80%",
-      "80% - 100%"
-    ];
-    const hexHash = {
-      "0% - 20%": "#FBEEEC",
-      "20% - 40%": "#f3c8c2",
-      "40% - 60%": "#e39787",
-      "60% - 80%": "#ce6751",
-      "80% - 100%": "#b3331d"
-    };
-
-    //    const steps = ["none", "#FBEEEC", "#f3c8c2", "#e39787", "#ce6751", "#b3331d"]
-    additionalSettings.afterElements = (
-      <HTMLLegend
-        valueHash={{}}
-        values={hexValues}
-        colorHash={hexHash}
-        colors={colors}
-      />
-    );
-  }
-  return {
-    xAccessor: metric1,
-    yAccessor: metric2,
-    axes: [
-      {
-        orient: "left",
-        ticks: 6,
-        label: metric2,
-        tickFormat: numeralFormatting
-      },
-      {
-        orient: "bottom",
-        ticks: 6,
-        label: metric1,
-        tickFormat: numeralFormatting
-      }
-    ],
-    points: !hexbin && data,
-    areas: hexbin && [{ coordinates: data }],
-    areaType: { type: "hexbin", bins: 10 },
-    areaStyle: (d: Object) => ({
-      fill: thresholds(d.percent),
-      stroke: "black"
-    }),
-    pointStyle: (d: Object) => ({
-      r: sizeScale(d[metric3]),
-      fill: colorHash[d[dim1]] || "black",
-      fillOpacity: 0.75,
-      stroke: "black",
-      strokeOpacity: 0.9
-    }),
-    hoverAnnotation: true,
-    responsiveWidth: false,
-    size: [height + 200, height + 50],
-    margin: { left: 75, bottom: 50, right: 150, top: 30 },
-    annotations: !hexbin && annotations,
-    annotationSettings: {
-      layout: { type: "marginalia", orient: "right", marginOffset: 30 }
-    },
-    tooltipContent: (hexbin && areaTooltip) || pointTooltip,
-    ...additionalSettings
-  };
-};
+import { semioticLineChart } from "./charts/line";
+import { semioticNetwork } from "./charts/network";
+import { semioticHierarchicalChart } from "./charts/hierarchical";
+import { semioticBarChart } from "./charts/bar";
+import { semioticScatterplot, semioticHexbin } from "./charts/xyplot";
+import { semioticSummaryChart } from "./charts/summary";
 
 const semioticParallelCoordinates = (
   data: Array<Object>,
