+++ packages/transform-dataresource/src/charts/xyplot.js
@@ -1,0 +1,243 @@
+/* @flow */
+import * as React from "react";
+
+import { sortByOrdinalRange } from "./shared";
+import { scaleLinear, scaleThreshold } from "d3-scale";
+import { numeralFormatting } from "../utilities";
+import HTMLLegend from "../HTMLLegend";
+
+const steps = ["none", "#FBEEEC", "#f3c8c2", "#e39787", "#ce6751", "#b3331d"];
+const thresholds = scaleThreshold()
+  .domain([0.01, 0.2, 0.4, 0.6, 0.8])
+  .range(steps);
+
+function combineTopAnnotations(
+  topQ: Array<Object>,
+  topSecondQ: Array<Object>,
+  dim2: string
+): any[] {
+  const combinedAnnotations = [];
+  const combinedHash = {};
+  [...topQ, ...topSecondQ].forEach(d => {
+    const hashD = combinedHash[d[dim2]];
+
+    if (hashD) {
+      const newCoordinates = (hashD.coordinates && [
+        ...hashD.coordinates,
+        d
+      ]) || [d, hashD];
+      Object.keys(combinedHash[d[dim2]]).forEach(k => {
+        delete combinedHash[d[dim2]][k];
+      });
+      combinedHash[d[dim2]].id = d[dim2];
+      combinedHash[d[dim2]].label = d[dim2];
+      combinedHash[d[dim2]].type = "react-annotation";
+      combinedHash[d[dim2]].coordinates = newCoordinates;
+    } else {
+      combinedHash[d[dim2]] = {
+        type: "react-annotation",
+        label: d[dim2],
+        id: d[dim2],
+        ...d
+      };
+      combinedAnnotations.push(combinedHash[d[dim2]]);
+    }
+  });
+  return combinedAnnotations;
+}
+
+export const semioticHexbin = (
+  data: Array<Object>,
+  schema: Object,
+  options: Object
+) => {
+  return semioticScatterplot(data, schema, options, true);
+};
+
+export const semioticScatterplot = (
+  data: Array<Object>,
+  schema: Object,
+  options: Object,
+  hexbin: boolean = false
+) => {
+  const height = options.height - 150 || 500;
+
+  const { chart, primaryKey, colors, setColor } = options;
+
+  const { dim1, dim2, metric1, metric2, metric3 } = chart;
+
+  const pointTooltip = (d: Object) => (
+    <div className="tooltip-content">
+      <h2>{primaryKey.map(p => d[p]).join(", ")}</h2>
+      {dim1 &&
+        dim1 !== "none" && (
+          <p>
+            {dim1}: {d[dim1]}
+          </p>
+        )}
+      <p>
+        {metric1}: {d[metric1]}
+      </p>
+      <p>
+        {metric2}: {d[metric2]}
+      </p>
+      {metric3 &&
+        metric3 !== "none" && (
+          <p>
+            {metric3}: {d[metric3]}
+          </p>
+        )}
+    </div>
+  );
+
+  const areaTooltip = (d: Object) => (
+    <div className="tooltip-content">
+      <h2
+        style={{
+          fontSize: "14px",
+          textTransform: "uppercase",
+          margin: "5px",
+          fontWeight: 900
+        }}
+      >
+        ID, {metric1}, {metric2}
+      </h2>
+      {d.binItems.map(d => (
+        <p
+          style={{
+            fontSize: "12px",
+            textTransform: "uppercase",
+            margin: "5px"
+          }}
+        >
+          {primaryKey.map(p => d[p]).join(", ")}, {d[metric1]}, {d[metric2]}
+        </p>
+      ))}
+    </div>
+  );
+
+  let sizeScale = e => 5;
+  const colorHash = { Other: "grey" };
+  const additionalSettings = {};
+
+  let annotations;
+
+  if (dim2 && dim2 !== "none") {
+    const topQ = [...data]
+      .sort((a, b) => b[metric1] - a[metric1])
+      .filter((d, i) => i < 3);
+    const topSecondQ = [...data]
+      .sort((a, b) => b[metric2] - a[metric2])
+      .filter(d => topQ.indexOf(d) === -1)
+      .filter((d, i) => i < 3);
+
+    annotations = combineTopAnnotations(topQ, topSecondQ, dim2);
+  }
+
+  if (metric3 && metric3 !== "none") {
+    const dataMin = Math.min(...data.map(d => d[metric3]));
+    const dataMax = Math.max(...data.map(d => d[metric3]));
+    sizeScale = scaleLinear()
+      .domain([dataMin, dataMax])
+      .range([2, 20]);
+  }
+  const sortedData = sortByOrdinalRange(
+    metric1,
+    (metric3 !== "none" && metric3) || metric2,
+    "none",
+    data
+  );
+
+  if (!hexbin && dim1 && dim1 !== "none") {
+    const uniqueValues = sortedData.reduce(
+      (p, c) =>
+        (!p.find(d => d === c[dim1].toString()) && [
+          ...p,
+          c[dim1].toString()
+        ]) ||
+        p,
+      []
+    );
+
+    uniqueValues.forEach((d, i) => {
+      colorHash[d] = i > 18 ? "grey" : colors[i % colors.length];
+    });
+
+    additionalSettings.afterElements = (
+      <HTMLLegend
+        valueHash={{}}
+        values={uniqueValues}
+        colorHash={colorHash}
+        setColor={setColor}
+        colors={colors}
+      />
+    );
+  } else if (hexbin) {
+    const hexValues = [
+      "0% - 20%",
+      "20% - 40%",
+      "40% - 60%",
+      "60% - 80%",
+      "80% - 100%"
+    ];
+    const hexHash = {
+      "0% - 20%": "#FBEEEC",
+      "20% - 40%": "#f3c8c2",
+      "40% - 60%": "#e39787",
+      "60% - 80%": "#ce6751",
+      "80% - 100%": "#b3331d"
+    };
+
+    //    const steps = ["none", "#FBEEEC", "#f3c8c2", "#e39787", "#ce6751", "#b3331d"]
+    additionalSettings.afterElements = (
+      <HTMLLegend
+        valueHash={{}}
+        values={hexValues}
+        colorHash={hexHash}
+        colors={colors}
+      />
+    );
+  }
+  return {
+    xAccessor: metric1,
+    yAccessor: metric2,
+    axes: [
+      {
+        orient: "left",
+        ticks: 6,
+        label: metric2,
+        tickFormat: numeralFormatting
+      },
+      {
+        orient: "bottom",
+        ticks: 6,
+        label: metric1,
+        tickFormat: numeralFormatting
+      }
+    ],
+    points: !hexbin && data,
+    areas: hexbin && [{ coordinates: data }],
+    areaType: { type: "hexbin", bins: 10 },
+    areaStyle: (d: Object) => ({
+      fill: thresholds(d.percent),
+      stroke: "black"
+    }),
+    pointStyle: (d: Object) => ({
+      r: sizeScale(d[metric3]),
+      fill: colorHash[d[dim1]] || "black",
+      fillOpacity: 0.75,
+      stroke: "black",
+      strokeOpacity: 0.9
+    }),
+    hoverAnnotation: true,
+    responsiveWidth: false,
+    size: [height + 200, height + 50],
+    margin: { left: 75, bottom: 50, right: 150, top: 30 },
+    annotations: !hexbin && annotations,
+    annotationSettings: {
+      layout: { type: "marginalia", orient: "right", marginOffset: 30 }
+    },
+    tooltipContent: (hexbin && areaTooltip) || pointTooltip,
+    ...additionalSettings
+  };
+};
